requirements.txt
discord.py==2.2.2
pyyaml==6.0.1
requests==2.32.3